requirements.txt
pyspark==3.5.0
kafka-python==2.0.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
psycopg2-binary==2.9.9
boto3==1.34.0